requirements.txt
Babel==2.3.4
Django==1.10.3
django-phonenumber-field==1.1.0
djangorestframework==3.5.3
phonenumberslite==7.7.3
Pillow==3.4.2
pytz==2016.7
